Organized ScAnalyzer GUI with executable script
Reorganized ScAnalyzer GUI to folder structure similar to GEECS_Scanner GUI.

Wrote a shell script which can be used as an effective executable to launch ScAnalyzer GUI. This includes base paths for both C and Z drives (although specific to HTU).
